requirements.txt
langchain==0.2.16
langchain-community==0.2.12
langchain-core==0.2.38
langchain-openai==0.1.22
chromadb==0.5.5
pandas==2.2.2
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.30.6
openai==1.43.0
